requirements.txt
openai>=1.0.0
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
pytz>=2023.3
rich>=13.0.0
asyncio-throttle>=1.0.0
colorama>=0.4.6
click>=8.1.0
